Adding Delete confirmation popup

+++ app/assets/javascripts/application.js
@@ -74,18 +74,21 @@ var addRow = function(){
 var removeRow = function(){
   $('#translationFields').on('click', '.remove', function(){
     // TODO Need to add Confirm/Canel button and conditions here.
-    alert('Are you sure you want to delete this translation?');
-    $.ajax({
-      url: '/translations/' + $(this).parent().parent().prev('input').val(),
-       type: 'DELETE',
-       success: function(response) {
-         // TODO Add success action here.
-         alert('Success!!!');
-       }
-    });
-     //alert($(this).parent().parent().prev('input').val());
-    $(this).parent().parent().remove();
+    var confirmation = confirm('Are you sure you want to delete this translation?');
+    if (confirmation == true) {
+      $.ajax({
+        // TODO The targeting here needs to be fixed.
+        url: '/translations/' + $(this).parent().parent().prev('input').val(),
+        type: 'DELETE',
+        success: function(response) {
+          // TODO Add success action here.
+          alert('Success!!!');
+        }
+      });
+      //alert($(this).parent().parent().prev('input').val());
+      $(this).parent().parent().remove();
       $('.languageSelect').selectunique('refresh');
+    }
   });
 }
 var ready = function(){
